+++ R/0_imports.R
@@ -66,3 +66,11 @@ hardhat::extract_preprocessor
 #' @importFrom hardhat extract_workflow
 #' @export
 hardhat::extract_workflow
+#'
+#' @importFrom hardhat extract_parameter_set_dials
+#' @export
+hardhat::extract_parameter_set_dials
+#'
+#' @importFrom hardhat extract_parameter_dials
+#' @export
+hardhat::extract_parameter_dials