+++ R/pull.R
@@ -41,7 +41,7 @@ pull_workflow_set_result <- function(x, id) {
 #' @export
 #' @rdname pull_workflow_set_result
 pull_workflow <- function(x, id) {
-  lifecycle::deprecate_soft("0.1.0", "pull_workflow()", "extract_worklfow()")
+  lifecycle::deprecate_soft("0.1.0", "pull_workflow()", "extract_workflow()")
   if (length(id) != 1) {
     rlang::abort("'id' should have a single value.")
   }